requirements.txt
streamlit
boto3
pdf2image
google-generativeai
pillow
